Gate MoE-LoRA kernels on triton availability
``kestrel/fused_moe`` is on the hot path for every moondream3 request, so
its module-level ``import triton`` in ``lora_kernels.py`` blocks all of
moondream3 on platforms where triton has no wheel (Windows, some Jetson
images). Make LoRA an opt-in feature:

- Split the two ``@triton.jit`` kernels out into ``_lora_triton_kernels``
  so only the LoRA code path pulls them in.
- Wrap ``import triton`` in lora_kernels in a try/except, set
  ``_TRITON_AVAILABLE`` flag. Module imports unconditionally.
- ``apply_moe_lora_batched`` and ``apply_moe_lora_single`` raise a clear
  RuntimeError at call time when triton is unavailable. Callers not using
  LoRA adapters never reach these.

Considered (and rejected) a plain-torch fallback: the kernels run inside
CUDA graph capture, and a torch equivalent would need fixed-shape /
no-host-sync operations plus masked scatters to stay capture-friendly
(~150 LoC of careful mask/gather plumbing). Gating is a cleaner first
step — adapters are unavailable on these platforms today but the base
model works. If plain torch LoRA becomes a hard requirement, the
capture-safe rewrite lives behind the same ``_TRITON_AVAILABLE`` switch.

Tested on Windows L4:
- ``from kestrel.fused_moe import FusedMoEModule`` succeeds without
  triton
- ``apply_moe_lora_batched`` raises the documented RuntimeError on call
- Moondream2 ChartQA smoke (10 samples, batch=4) matches pre-change
  numbers: 10.54 req/s, 357 ms P50.
